% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collocation_functions.R
\name{col_network}
\alias{col_network}
\title{Create a table for plotting a collocational network}
\usage{
col_network(col_1, ...)
}
\arguments{
\item{col_1}{A collocations object produced by collocates_by_MI().}

\item{...}{Other collocations objects for plotting.}
}
\value{
A tidygraph table for network plotting.
}
\description{
This function operationalizes the idea of collcational networks described by
Brezina, McEnery & Wattam (2015):
https://www.jbe-platform.com/content/journals/10.1075/ijcl.20.2.01bre
The function takes data.frames produced by the collocates_by_MI() function above
and generates a tidygraph data object for plotting in ggraph.
}
