% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_functions.R
\name{dispersions_all}
\alias{dispersions_all}
\title{Dispersion measures for all tokens in a corpus}
\usage{
dispersions_all(target_dfm)
}
\arguments{
\item{target_dfm}{The target document-feature matrix.}
}
\value{
A data.frame containing dispersion measures for the tokens in the document-feature matrix.
}
\description{
The dispersions_all() function calculates a subset of of the most common dispersion measures
for all of the tokens in a document-feature matrix and returns a data.frame.
For example: dispersions_all(my_dfm)
}
