% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_functions.R
\name{dispersions_token}
\alias{dispersions_token}
\title{Dispersion measures for a token}
\usage{
dispersions_token(target_dfm, token)
}
\arguments{
\item{target_dfm}{The target document-feature matrix.}

\item{token}{The token for which dispersion measures are to be calculated.}
}
\value{
A named list of dispersion measures.
}
\description{
The dispersions_token() function calculates
the dispersion measures for a single token.
For example: dispersions_token(my_dfm, "cat")
It returns a named list with all of the
dispersion measures discussed by S.T. Gries:
(http://www.stgries.info/research/dispersion/links.html)
}
