% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ds_dict}
\alias{ds_dict}
\title{Simplified DocuScope dictionary}
\format{
quanteda dictionary object (see \code{quanteda::dictionary()})
}
\usage{
ds_dict
}
\description{
A quanteda dictionary (see \code{quanteda::dictionary()}) that allows for token
classification via lookups, using \code{quanteda::tokens_lookup()}. This is a
highly simplified version of
\href{https://docuscospacy.readthedocs.io/en/latest/docuscope.html}{DocuScope}.
You can find the \href{https://docuscospacy.readthedocs.io/en/latest/docuscope.html#categories}{category descriptions here}.
}
\details{
To use the dictionary, fixed matches should be used and tokens should be
prepared with minimal processing:

\if{html}{\out{<div class="sourceCode">}}\preformatted{tokens(some_corpus, remove_punct = FALSE, remove_numbers = FALSE,
       remove_symbols = FALSE, what = "word")
}\if{html}{\out{</div>}}

DocuScope follows similar principles to a \href{https://www.saifmohammad.com/WebPages/NRC-Emotion-Lexicon.htm}{Sentiment Lexicon} or
\href{https://lit.eecs.umich.edu/geoliwc/liwc_dictionary.html}{LIWC} (Linguistic
Inquiry and Word Count), but is orders of manitude larger and organizes its
categories according to a rhetorical orientation to language.
}
\examples{
library(quanteda)

corpus <- sample_corpus$text
names(corpus) <- sample_corpus$doc_id

toks <- tokens(corpus, remove_punct = FALSE, remove_numbers = FALSE,
               remove_symbols = FALSE, what = "word")

tokens_lookup(toks, ds_dict, levels = 1, valuetype = "fixed")
}
\keyword{datasets}
