% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{excel_style}
\alias{excel_style}
\title{A function for expanding letter sequences.}
\usage{
excel_style(i)
}
\arguments{
\item{i}{Index of alphabetic character}
}
\value{
A vector of character combinations in the style of Excel column headers
}
\description{
A function for expanding letter sequences.
}
