% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extend}
\alias{extend}
\title{A function for expanding letter sequences.}
\usage{
extend(alphabet)
}
\arguments{
\item{alphabet}{An alphabet to be expanded}
}
\description{
A function for expanding letter sequences.
}
