% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_functions.R
\name{frequency_table}
\alias{frequency_table}
\title{Descriptive measures for all tokens in a corpus}
\usage{
frequency_table(target_tkns)
}
\arguments{
\item{target_tkns}{The target quanteda tokens object.}
}
\value{
A data.frame containing absolute frequency, relative frequency, average reduced frequency, and deviation of proportions.
}
\description{
The frequency_table() function aggregates useful descriptive measures: absolute frequency, relative frequency, average reduced frequency, and deviation of proportions.
}
