% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyness_functions.R
\name{keyness_pairs}
\alias{keyness_pairs}
\title{Pairwise keyness values from any number of dfms}
\usage{
keyness_pairs(dfm_a, dfm_b, ..., yates = FALSE)
}
\arguments{
\item{dfm_a}{A document-feature matrix}

\item{dfm_b}{A document-feature matrix}

\item{...}{Additional document-feature matrices}

\item{yates}{A logical value indicating whether the "Yates" correction should be performed}
}
\value{
A data.frame containing pairwise keyness comparisons of all dfms
}
\description{
This function takes any number of quanteda dfm objects and returns a table of log-likelihood values, effect sizes
using Hardie's log ratio and p-values
}
