% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiword_expressions}
\alias{multiword_expressions}
\title{List of common multi-word expressions}
\format{
Character vector of multi-word expressions
}
\usage{
multiword_expressions
}
\description{
Multi-word expressions are common English phrases, like "a lot" or
"for keeps", that we may want to count as a unit rather than counting the
words individually.
}
\keyword{datasets}
