% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{normalizing_factor}
\alias{normalizing_factor}
\title{A function for detecting the size of a corpus and setting the narmalizing factor to the nearest power of 10.}
\usage{
normalizing_factor(x)
}
\arguments{
\item{corpus_total}{The total number of words in the corpus}
}
\value{
A named vector
}
\description{
A function for detecting the size of a corpus and setting the narmalizing factor to the nearest power of 10.
}
