% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{preprocess_text}
\alias{preprocess_text}
\title{Pre-process texts}
\usage{
preprocess_text(
  txt,
  contractions = TRUE,
  hypens = TRUE,
  punctuation = TRUE,
  lower_case = TRUE,
  accent_replace = TRUE,
  remove_numbers = FALSE
)
}
\arguments{
\item{txt}{A character vector}

\item{contractions}{A logical value to separate contractions into two tokens}

\item{hypens}{A logical value to separate hypenated words into two tokens}

\item{punctuation}{A logical value to remove punctuation}

\item{lower_case}{A logical value to make all tokens lower case}

\item{accent_replace}{A logical value to replace accented characters
with un-accented ones}

\item{remove_numbers}{A logical value to remove numbers}
}
\value{
A character vector
}
\description{
A simple function that requires a readtext object.
It then processes the text column using basic regex substitutions.
The default is to add a space before possessives and contractions.
This will force their tokenization in quanteda.
So that "Shakespeare's" will be counted
as two tokens rather than a single one.
It is easy to add or delete substations as fits your analytical needs.
}
