% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{readtext_lite}
\alias{readtext_lite}
\title{Read texts from a vector of paths}
\usage{
readtext_lite(paths)
}
\arguments{
\item{paths}{A vector of paths to text files that are to be read in.}
}
\value{
A readtext data.frame
}
\description{
Replaces the readtext::readtext function that
reads a lists of text files into a data frame.
}
