% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_corpus}
\alias{sample_corpus}
\title{Sample text corpus}
\format{
\describe{
\item{doc_id}{Document ID, in the form type_number.}
\item{text}{Document text.}
}
}
\usage{
sample_corpus
}
\description{
A small corpus compiled following the sampling procedures of the \href{https://www.english-corpora.org/coca/}{Corpus of Contemporary American English}.
However, it contains only 50 texts from each type, and each text is only
about 2,500 words. Thus, it is similar to the \href{https://www1.essex.ac.uk/linguistics/external/clmt/w3c/corpus_ling/content/corpora/list/private/brown/brown.html}{Brown family of corpora}
in its size (roughly 1 million words).
}
\details{
The corpus contains data from 8 text-types:
\itemize{
\item Academic
\item Blog
\item Fiction
\item Magazine
\item News
\item Spoken
\item Television & Movie
\item Web
}
}
\section{Warning}{
These data are included only for demonstration purposes. They not compiled to
be used for research.
}

\keyword{datasets}
