% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_functions.R
\name{word_list}
\alias{word_list}
\title{Descriptive measures for all tokens in a corpus}
\usage{
word_list(target_tkns)
}
\arguments{
\item{target_tkns}{The target quanteda tokens object.}
}
\value{
A data.frame containing absolute frequency, relative frequency, average reduced frequency, and deviation of proportions.
}
\description{
The word_list() function aggregates useful descriptive measures: absolute frequency, relative frequency, average reduced frequency, and deviation of proportions.
}
