% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collocation_functions.R
\name{col_network}
\alias{col_network}
\title{Create a table for plotting a collocational network}
\usage{
col_network(col_1, ...)
}
\arguments{
\item{col_1}{A collocations object produced by collocates_by_MI().}

\item{...}{Other collocations objects for plotting.}
}
\value{
A tidygraph table for network plotting.
}
\description{
This function operationalizes the idea of collcational networks described by
Brezina, McEnery & Wattam (2015). The function takes data.frames produced by
\code{collocates_by_MI()} and generates a tidygraph data object for plotting in
ggraph.
}
\examples{
\donttest{
library(quanteda)
texts <- c("The big cat sat on the mat",
           "The big cat ran very fast",
           "The big cat likes to hunt mice",
           "Time flies like an arrow",
           "Money may be important, but so is time")
toks <- tokens(texts)
# Create collocation objects for different node words
cat_collocations <- collocates_by_MI(toks, "cat", left = 1, right = 1)
time_collocations <- collocates_by_MI(toks, "time", left = 1, right = 1)
# Create network object
network <- col_network(cat_collocations, time_collocations)
print(network)
}
}
\references{
Brezina, McEnery and Wattam (2015).
"Collocations in context: A new perspective on collocation networks."
\emph{International Journal of Corpus Linguistics} 20 (2), 139-173.
\doi{10.1075/ijcl.20.2.01bre}
}
