% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{excel_style}
\alias{excel_style}
\title{A function for creating Excel-style column headers}
\usage{
excel_style(i)
}
\value{
A sequence of characters.
}
\description{
A function for creating Excel-style column headers
}
