% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyness_functions.R
\name{keyness_table}
\alias{keyness_table}
\title{Keyness measures for all tokens in a corpus}
\usage{
keyness_table(target_dfm, reference_dfm, yates = FALSE)
}
\arguments{
\item{target_dfm}{The target document-feature matrix}

\item{reference_dfm}{The reference document-feature matrix}

\item{yates}{A logical value indicating whether the "Yates" correction should be performed}
}
\value{
A data.frame containing the log-likelihood, log ratio, absolute frequencies, relative frequencies, and dispersions
}
\description{
The keyness_table() function returns the log-likelihood of the target vs. reference corpus, effect sizes by log ratio, p-values, absolute frequencies, relative frequencies, and deviation of proportions.
}
