% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyness_functions.R
\name{keyness_table}
\alias{keyness_table}
\title{Descriptive measures for all tokens in a corpus}
\usage{
keyness_table(target_dfm, reference_dfm, yates = FALSE)
}
\arguments{
\item{target_dfm}{The target document-feature matrix}

\item{reference_dfm}{The reference document-feature matrix}

\item{yates}{A logical value indicating whether the "Yates" correction should be performed}
}
\value{
A data.frame containing the log-likelihood, log ratio, absolute frequencies, relative frequencies, and dispersions
}
\description{
The frequency_table() function aggregates useful descriptive measures: absolute frequency, relative frequency, average reduced frequency, and deviation of proportions.
}
