% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyness_functions.R
\name{log_like}
\alias{log_like}
\title{Log-likelihood calculation}
\usage{
log_like(n_target, n_reference, total_target, total_reference, correct = FALSE)
}
\arguments{
\item{n_target}{The non-normalized token count in the target corpus}

\item{n_reference}{The non-normalized token count in the reference corpus}

\item{total_target}{The total number of tokens in the target corpus}

\item{total_reference}{The total number of tokens in the reference corpus}

\item{correct}{Whether to perform the Yates correction}
}
\value{
A numeric value representing log-likelihood
}
\description{
Log-likelihood tests the frequencies of tokens in one corpus vs. another. It
is often used instead of a chi-square test, as it has been shown to be more
resistant to corpora of varying sizes. For more detail see:
\url{http://ucrel.lancs.ac.uk/llwizard.html}
}
